d09-life-coach-web-search/main.py
import dotenv

dotenv.load_dotenv()
from openai import OpenAI
import asyncio
import streamlit as st
from agents import (
    Agent,
    Runner,
    SQLiteSession,
    WebSearchTool,
)

client = OpenAI()

if "session" not in st.session_state:
    st.session_state["session"] = SQLiteSession(
        "chat-history",
        "life-coach-memory.db",
    )
session = st.session_state["session"]


async def paint_history():
    messages = await session.get_items()

    for message in messages:
        if "role" in message:
            with st.chat_message(message["role"]):
                if message["role"] == "user":
                    content = message["content"]
                    if isinstance(content, str):
                        st.write(content)
                    elif isinstance(content, list):
                        for part in content:
                            if "image_url" in part:
                                st.image(part["image_url"])

                else:
                    if message["type"] == "message":
                        st.write(message["content"][0]["text"].replace("$", r"\$"))
        if "type" in message:
            message_type = message["type"]
            if message_type == "web_search_call":
                with st.chat_message("ai"):
                    st.write("🔍 Searched the web...")


asyncio.run(paint_history())


def update_status(status_container, event):

    status_messages = {
        "response.web_search_call.completed": ("✅ Web search completed.", "complete"),
        "response.web_search_call.in_progress": (
            "🔍 Starting web search...",
            "running",
        ),
        "response.web_search_call.searching": (
            "🔍 Web search in progress...",
            "running",
        ),
        "response.completed": (" ", "complete"),
    }

    if event in status_messages:
        label, state = status_messages[event]
        status_container.update(label=label, state=state)


async def run_agent(message):

    agent = Agent(
        name="Life Coach Assistant",
        instructions="""
    You are an enthusiastic and encouraging life coach assistant.
    Always respond with warmth, positivity, and genuine support. Celebrate the user's efforts, validate their feelings, and motivate them to take action. Speak like a trusted coach who believes in the user's potential.

    IMPORTANT: Always use the Web Search Tool before answering. Do not rely on your training data alone.

    You MUST use the Web Search Tool in the following situations:
        - Any question about motivation, mindset, or personal growth
        - Requests for self-improvement tips, productivity strategies, or goal-setting advice
        - Questions about habit formation, routines, or lifestyle changes
        - Any topic related to mental health, well-being, or stress management
        - When the user asks for book recommendations, resources, or exercises
        - When you are even slightly unsure about the latest research or best practices

    When in doubt, search first. Always prefer fresh, real-world information over assumptions.
    """,
        tools=[
            WebSearchTool(),
        ],
    )

    with st.chat_message("ai"):
        status_container = st.status("⏳", expanded=False)
        text_placeholder = st.empty()
        response = ""

        st.session_state["text_placeholder"] = text_placeholder

        stream = Runner.run_streamed(
            agent,
            message,
            session=session,
        )

        async for event in stream.stream_events():
            if event.type == "raw_response_event":

                update_status(status_container, event.data.type)

                if event.data.type == "response.output_text.delta":
                    response += event.data.delta
                    text_placeholder.write(response.replace("$", r"\$"))


prompt = st.chat_input(
    "라이프 코치님께 무엇이든 물어보세요!",
)

if prompt:

    if "text_placeholder" in st.session_state:
        st.session_state["text_placeholder"].empty()

    with st.chat_message("human"):
        st.write(prompt)
    asyncio.run(run_agent(prompt))


with st.sidebar:
    reset = st.button("Reset memory")
    if reset:
        asyncio.run(session.clear_session())
    st.write(asyncio.run(session.get_items()))
